Add pnpm lock file
I can't figure out how to make rules_nodejs and rules_js live happily
alongside one another during the migration. So it'll be a rather big
patch when the switchover actually happens. For the sake of keeping
that switchover patch as small as possible, this patch here adds the
pnpm lock file that we'll be using with rules_js.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: I74f2a5d94b5eec096001b2efd68192fbdf564e51