Fix aos_dump/aos_send autocomplete w/ zsh
This change add an additional flag to aos_cli_utils which disables
returning <channel> <message type> pairs. I spent a while digging
into the zsh auto-completion subsystem to get it or aos to handle
the pairs correctly and was not able to find a good way to get it
working.

I do also think this gives a slightly more managable auto-complete
list at the command line.

Change-Id: Ia79dddb68db6fe930d5944ad2ee858a1e8d41972
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>