Disable lockless_queue_test temporarily
In order to cleanly apply the rules_js patch, I need to purge the
caches of all the workers. To do that, I need to take them offline,
purge the cache, bring them back and re-validate the rules_js patch.
This will be much faster when the lockless_queue_test isn't running.

This patch disables the test. After the migration is complete, the
test will be re-enabled.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: Iabbc7fd62de190b2ac01e3e0410b2f9bda74444f