Refactor some code for scraping thebluealliance.com in the background
I want to make it so that the match list is imported automatically.
Right now we have to manually refresh the match list as elimination
matches happen.

This patch refactors the ranking scraping code so that we can add a
match list scraper easily in a future patch.

I made a few things generic so that it'll be easier to reuse the code
for more things later.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: Ia8164b62deceddeda683ac3080e99e0fe1b5904a