Fix errors in //scouting/deploy
In I24e1687732eac36561cb8f1cfb4fb1a517895704 I added the `--clear-db`
option to the script. Turns out I didn't test it enough. The commands
work when typed on the console, but I made some errors when adding
them to the script.

The first fix is to set `shell=True`. Otherwise, the whole command
is interpreted as an executable.

The second fix is to quote the whole command passed to ssh. Otherwise,
the command doesn't get interpreted properly. Not exactly sure why.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: If316847874ce61adeb05c230b711663bfbe6fd0e