Fix drivetrain randomly freezing
Omar noticed that there was a neutral command fighting the control
command.  Neither of us know why, so we are no longer doing anything
with neutral...  We ran 100 reboots without the diag server and 100 with
and it worked in both cases.

He's going to file an internal ticket to debug it on Monday.  This
should get us through the rest of the season.

Change-Id: Ib922501cfe2cd1ce4d6933054920b6a63b2420e8
Signed-off-by: Austin Schuh <austin.linux@gmail.com>