Fix flaky //scouting/webserver/requests/debug/cli:cli_test
We're seeing timeouts in this test. After adding some debugging, it
looks like it's fixed by making stronger guarantees about the startup
order. Every once in a while the scouting server started up faster
than the fake TBA server and fail to scrape the data. Since we only
scrape the data every 10 minutes, the test times out before scraping
completes.

The fix here is to start the fake TBA server before the webserver.
That lets the webserver scrape successfully on startup.

I left the additional debugging in place because it might be helpful
in the future.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: Ibe32b0e622319ff4813c056054aea8543e4ae2e2