Updating the extrinsic calibration on 971-1 and 9971-2
I was looking at the calibrations, and most seem good except 971-1 and 9971-2.
By good, I mean they are reasonably close to CAD/expected values.

I noticed that the 971-1 images had a significant outlier in the camera poses,
so I manually removed that one and it looks much more like I expected it to.

We recaptured the calibration on 9971-2, and it converged better, and it returned
values that were much more reasonable, esp. in x

Change-Id: I428a9606da115291f23f244f8a66383607fe8e01
Signed-off-by: Jim Ostrowski <yimmy13@gmail.com>