Parametrize the DriverRank.jl script
This patch makes the `DriverRank.jl` accept two arguments. The first
is the CSV file to read the input data from. The second is the CSV
file to save the parsed results to.

This patch also updates the Go code parsing the output CSV because it
contains a "header". I.e. labels for the columns.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: Iaeeb02316bc04fbf27962f0f8d9ac12880a61b8d