Reduce amount of stuff forwarded to/from the roborio
Nothing needs the drivetrain Position message, don't forward the
drivetrain Output to the "logger" pi, and reduce the rate of the
LocalizerOutput (since I don't see a strong need to run it at 200
Hz...).

Note that we are still forwarding starter RPC and Status messages
between the roborio and everything. This isn't strictly necessary, but
is useful enough that I want to keep it for now.

Change-Id: I44a9b20a065bb2005d425f4f6d1676e06c2d3d3d
Signed-off-by: James Kuszmaul <jabukuszmaul@gmail.com>