scouting: Let web page know which matches have been scouted
One of the big complaints is that the match list on the scouting app
is annoying to scroll through as the day goes on. One approach for
addressing this is to hide the matches that have already been scouted.
A simple solution for this is to enhance the request for the match
list with some booleans. Each boolean says whether that particular
team has already been scouted.

Change-Id: I3801958edf09d2edca01dbb8d54fc12cb65e52c4
Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>