scouting: Add database helper to find status for specific match
I am working on letting the scouting app hide matches that are already
scouted. The approach that I'm currently going with involves querying
the database for an individual team's match. This patch adds the
querying functionality.

A future patch will enhance the logic in `requests.go` to add an
additional flag to `RequestAllMatchesResponse`. This flag will tell
the users whether a specific robot in a match has already been
scouted.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: I7365aee212bc304ab935c7a4c95e5783008ab133