Enhance scouting server to pick up local `scouting_config.json`
I'm tired of manually specifying the config on the command line. It's
simpler to set the default such that it picks up the
`scouting_config.json` file from the local directory.

I also took this opportunity to enhance the docs for the flag. The
`year` and `event_code` fields are pretty much required now.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: I3687d6d77e3433aaad0918904f83901c988cccdd