Make jinja2_template properly declare its output
Previously, the `name` attribute was being used for the jinja output,
however that attribute one is a default bazel attribute that is
not marked as an output. This was a weird bazel edge case which was ok
to be used in genrules, but the generated files could not be accessed in
a py_binary because the outputs weren't declared correctly.

Add a separate `out` field that is explicitly declared as an output.
To keep the API call the same, make a wrapper macro that fills this
field out, given a name parameter.

Additionally, add a `list_parameters` field to the build rule.

Change-Id: I2aefb7b6f552f81563bd78d3b909587d49ac83db
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>