scouting: Skip team selection after clicking on a team in the match list
It's not really useful to have the scouts go through the Team
Selection screen. The information is known to be correct. With this
patch, the user is automatically brought to the 'Init' screen when
they click on a team in the match list.

I also took this opportunity to clean up some stale signals/events
that weren't being used.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: I2da4178465c217664c21251ffc1db46c42ca7445