Fix arm freakout
There was an uninitialized variable which James found.  Manually setting
it to a bad value triggers the arm to freak out when disabled, and the
code below didn't clear it.  I confirmed that initializing it to 0 fixes
it...

I have no clue how to write a test for this reasonably.  We'd need
sanitizers working.

Change-Id: I53881cf2e83ea9e610e8ac86de40ec8c4cdd3bb1
Signed-off-by: Austin Schuh <austin.linux@gmail.com>