Scouting: Improve entry spacing for better visibility
This change dynamically changes spacing between buttons so that
all buttons and elements are visible without scrolling on most
devices. Currently, the switch to teleop button is hidden on
many devices causing scouts to forget to switch to teleop.
Also, this change makes the switch to teleop button color
black so it is more visible.

Signed-off-by: Filip Kujawa <filip.j.kujawa@gmail.com>
Change-Id: I5c1772eaff9ddd96278abea7d04e16e203cd611a