Add statistics and filter outliers in mapping
Compute the standard deviation and error for each position and rotation
element in our constraints, so we can understand how good the resulting
solved target map will be. Then, filter based on these stats to get rid
of constraints with large errors, and resolve for a better target map.

Also, delete the constraint seeding because these get flagged as
outliers.

Signed-off-by: milind-u <milind.upadhyay@gmail.com>
Change-Id: I364311f8a4558eaf259882488b8a7ff6d5079b09