scouting: Convert actions to stats when submitted
Now that actions are submitted, we need to convert them into stats.
This patch does that by calling the correct conversion function and
then inserting the result into the database.

I had to adjust the scouting_test slightly because the behaviour
changed slightly. The test doesn't clear the database between test
cases. So the one test case submitting scout data means it can't be
scouted again. The button is now disabled.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: Ib3e68103c8064c61e628711c9fe375919372e18c