scouting: Fix docking/engaged data collection during auto
Looks like in a recent patch, we violated the scope of the template
variables that represent the checkboxes for docking, engaged, etc.
https://angular.io/guide/template-reference-variables#template-variable-scope

I.e. the button to submit the data could not access the checkboxes
because it was in its own, distinct `ngIf` scope.

This patch fixes the issue by putting the buttons in the same scope as
the checkboxes.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: I30ea8c3435b60dc267541c0178b6dbe50e52c0e6