Add a Primal-dual interior-point method solver
This sets us up to solve nonlinear, convex problems efficiently.  I'm
not 100% sure this has the right interfaces, but it appears to work for
a toy problem and is worth checking in.

This is from https://web.stanford.edu/~boyd/cvxbook/bv_cvxbook.pdf
section 11.7.  Parts of the code are from the multinode timestamp
solver, though that solver has enough workarounds for the problem that
is isn't easily generalized.

Change-Id: I38a727b81df3ebfc24a42a633ab1cf74cf5ac692
Signed-off-by: Austin Schuh <austin.linux@gmail.com>