Maybe fix zero sized messages
We are seeing occasional zero sized
messages.  The only place I can guess these are coming from is a
"partial" buffer from the lzma encoder which is actually 0 length.  Add
in both more debugging, and code to potentially fix that cause.  We'll
want to revert parts of this patch if this proves not to be the issue
(and make it a CHECK).

F20230502 20:28:21.927556  5914 log_backend.cc:65] Check failed: size > 0 (0 vs. 0) : Nobody should be sending empty messages.

Change-Id: I78609e20d086e43f02a67510c6dd1461ab77ecb9
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>