Make SendingAfterSendingTooFast less sensitive to timing
We were trying to hit a 1ms loop and not miss more than 200 in a second.
That's a tall ask for a VM server in the cloud.  Instead, lower the
frequency and just send more messages each cycle.  Also, if we get
behind by a cycle or two, send out more messages to compensate...

Change-Id: I6777842e86ab3d79b46267c40c4a05097baf3d09
Signed-off-by: Austin Schuh <austin.linux@gmail.com>