Strip down message size for MessageBridgeServer
This makes it much much less costly to have a pool of messages we use
for receiving data into on the server side.  The server only really
receives the connection message along with any remote timestamps being
sent back.

While we are here, make sure the sha256 is the right length...  We were
blindly treating it as a string, and that causes it to overflow the
destination buffer.

Change-Id: Id4bfa0d1d9c25e0beead39248f32b361b722059d
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>