Use remote index to seed last_queue_index in message_bridge_client
We were using the local queue index, for some reason.

This would have resulted in potentially duplicated messages getting sent
by the message bridge client if it ever restarted. However, it would
realistically only ever do so with reliable messages. And with reliable
messages, if the local and remote queue indices were perfectly in sync
then it would not have been an issue.

Can probably write a test if needed; the sequencing to set it up is
mildly obnoxious.

Change-Id: I742d469431eb5d683bf943f6e83efd8557ca2e41
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>