Size WMem large enough to hold all published messages at once
If the server sends a message to all the clients at once, it overflows
the kernel tx queue for message bridge.  We used to account for things
poorly enough that it would mostly work.  We now are much tighter about
our needs, so we don't.  Track the memory required more explicitly here.

Change-Id: I1c9430e7f2424961132bc3f786a9af1c426650ad
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>