Initialize channel memory in parallel
When there are a large amount of channels, it takes a while to map them
and make sure there are pages behind them and owned by starterd.  This
turns out to be many seconds on larger systems.  This can be done in
parallel, and achieves a 50+% speedup on that part of the startup
process.

Change-Id: Id45e279b3dcd19e7d81c3b5b75ef324519f3721a
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>