Fix a logger issue when disk is full on startup
There's an null dereference in the code when the logger fails to
create files.

This is happening because of this line of code in
`third_party/aos/aos/events/logging/logfile_utils.h` line 109:

  WriteStats *WriteStatistics() const { return log_sink_->WriteStatistics(); }

The `log_sink_` pointer doesn't point to anything when the
`DetachedBufferWriter` is constructed with the special
`already_out_of_space_t` marker.

This patch fixes that by instantiating a dummy `OutOfDiskSpaceLogSink`
implementation of `LogSink` when `DetachedBufferWriter` is constructed
with the special `already_out_of_space_t` marker. This dummy
implementation refuses to do anything useful and just reports back
that the disk is full.

I also took this opportunity to mark the `DetachedBufferWriter`
destructor as `virtual`. This is imperative because it's inherited
from in at least one place.

Change-Id: I780dcd5a9418f11216da7cc96e3685e35190e998
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>