Remove redundant ignore_missing_messages in Register()
Two pieces of code were doing the same thing (ensuring that we don't
complain about mismatched timestamps/data before startup), and one
was only being used in StartAfterRegister() (and had been added earlier).
Remove the redundant check, and make StartAfterRegister() simpler.

This exposed an error in the multinode_logger_test which was incorrectly
ignoring missing message data that occurred between the starts
of a pi1 and pi2 logger due to not reading a log part.

Take the opportunity to add some docs on what "start time" means in the
logger, since I needed to satisfy myself that we were using the start
time correctly in order to be willing to make this change.

Change-Id: I2c4dc98ff5646ecd1f8e85f8ca01e1ad178d3b61
Signed-off-by: James Kuszmaul <jabukuszmaul+collab@gmail.com>