Add LogReplayerStats message
This change adds start and stop times of when replay happens in
log_replayer and the replay_config used as an fbs message that is sent
on the /replay channel when replay is complete.
The start and stop times are needed to properly determine
when exactly log replay was happening to analyze the results of a
resulting diagnostic log only within the context of replay.

Change-Id: Ic86c41f11689597986e02b6154f18bb5c8552b2c
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>