Add AOS docs covering ShmEventLoop & multi-node constructs
Add a few thousand words covering the ShmEventLoop and the basics of how
multi-node systems work.

Still a fair amount to be written on multi-node systems overall.

This also updates/adds to some of the example code we have:

* Add options that allow for experimenting with higher rates & fetchers
  in ping/pong (since I talked about performance benefits of watchers vs
  fetchers in the docs).
* Update the starter_demo script to no longer call `aos_starter`
  `starter_cmd` and to provide the `aos_timing_report_streamer`.
* Add sample code for using the `ServerStatistics` clock offsets.

This change is primarily meant to put factually accurate and clear
information into the docs, and less so about getting perfectly styled
prose.

Change-Id: Ieff45b0ef45e3390e0f98630a65651028206a9f0
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>