aos: Let log_stats find inefficiently packed flatbuffer messages
James Kuszmaul found that some messages are inefficiently packed. This
was causing unnecessary CPU usage increases of the logger.

In order to find any other channels with inefficiently packed
messages, this patch adds a new `--print_repack_size_diffs` flag. This
flag prints the number of bytes saved when messages on the
corresponding channels are naively repacked. The repacking method is
currently _very_ naive. Only use it when necessary. Otherwise, it'll
take a long time to analyze your log. I recommend using it with the
`--run_for` flag.

Change-Id: I0e18acccddb3d40af138559b40d356f22960b149
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>