Make sure OnEnd is always called.
OnEnd was not being called if a log file had missing data and terminated
early.  A lot of applications like to use OnEnd to be notified that a
log has ended, and to trigger cleanup/writing to disk behavior.  This
seems like a reasonable assumption worth supporting.

Add a test to confirm this behavior explicitly, then add the check to
every test (number of starts == number of stops), then fix it.

Change-Id: I1b0aeeb963f813c83a625690827034394a2c379c
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>