Fix message_bridge_auth_client crash
When we lose connection to the message bridge authentication server
(maybe from a reboot or network going down), the client's watcher
can block for long until we get the gRPC response back from the server.

This leads to 2 follow-up crashes.

1. We crash because the watcher tries to read the next message but it's
long gone (because we blocked for so long)
2. We restart and we have a full queue of messages to handle which we do
as soon as possible and crash because we sent too many messages too fast.

This change adds deadline to the gRPC request which helps prevent the
bottle-up. Additionally, it switches the request/response model of
message bridge authentication client to a polling model which effectively
rates-limit how fast we can send the response back.

Change-Id: I3bdfabd61bae76ecba2e663245763289ed479d45
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>