Clear the point cache in NoncausalTimestampFilter's pointer
When we change the cached t0, t1, there was a code path which would end
up leaving the old, bad points from before.  We would then interpolate
very poorly, failing line search and making the solver not converge.

Add a check to catch the bad code path, and then fix it.

Change-Id: Ie544a47c704ee009f0dde2d4c22907f4ffd7189f
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>